requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Object Tracking
filterpy>=1.4.5
scipy>=1.10.0
scikit-learn>=1.3.0

# GUI Framework
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
pyqtgraph>=0.13.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0

# Database
sqlite3-python

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Logging and Monitoring
loguru>=0.7.0

# Notifications
requests>=2.31.0

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-qt>=4.2.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# Performance
numba>=0.57.0
